+++ src/main/java/com/breakinblocks/nautec/content/bacteria/SimpleBacteria.java
@@ -14,7 +14,7 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 
-public record SimpleBacteria(LongRange initialSize, Resource.ItemResource resource, BacteriaStats<?> stats) implements Bacteria {
+public record SimpleBacteria(LongRange initialSize, Resource resource, BacteriaStats<?> stats) implements Bacteria {
     public static Builder of() {
         return new Builder();
     }
@@ -33,13 +33,13 @@ public record SimpleBacteria(LongRange initialSize, Resource.ItemResource resour
         public static final Serializer INSTANCE = new Serializer();
         public static final MapCodec<SimpleBacteria> MAP_CODEC = RecordCodecBuilder.mapCodec(instance -> instance.group(
                 LongRange.MAP_CODEC.fieldOf("initial_size").forGetter(SimpleBacteria::initialSize),
-                Resource.ItemResource.CODEC.fieldOf("bacteria").forGetter(SimpleBacteria::resource),
+                Resource.CODEC.fieldOf("bacteria").forGetter(SimpleBacteria::resource),
                 BacteriaStats.CODEC.fieldOf("stats").forGetter(SimpleBacteria::stats)
         ).apply(instance, SimpleBacteria::new));
         public static final StreamCodec<RegistryFriendlyByteBuf, SimpleBacteria> STREAM_CODEC = StreamCodec.composite(
                 LongRange.STREAM_CODEC,
                 SimpleBacteria::initialSize,
-                Resource.ItemResource.STREAM_CODEC,
+                Resource.STREAM_CODEC,
                 SimpleBacteria::resource,
                 BacteriaStats.STREAM_CODEC,
                 SimpleBacteria::stats,
@@ -62,7 +62,7 @@ public record SimpleBacteria(LongRange initialSize, Resource.ItemResource resour
 
     public static class Builder implements Bacteria.Builder<SimpleBacteria> {
         private LongRange initialSize = LongRange.of(0, 0);
-        private Resource.ItemResource resource = new Resource.ItemResource(Items.AIR);
+        private Resource resource = Resource.EMPTY;
         private FloatRange growthRate = FloatRange.of(0F, 0F);
         private FloatRange mutationResistance = FloatRange.of(0F, 0F);
         private FloatRange productionRate = FloatRange.of(0F, 0F);
@@ -76,6 +76,11 @@ public record SimpleBacteria(LongRange initialSize, Resource.ItemResource resour
 
         public Builder resource(Item resource) {
             this.resource = new Resource.ItemResource(resource);
+            return this;
+        }
+
+        public Builder resource(Resource resource) {
+            this.resource = resource;
             return this;
         }