+++ src/main/java/com/breakinblocks/nautec/datagen/DataGatherer.java
@@ -49,7 +49,7 @@ public class DataGatherer {
         DatapackRegistryProvider datapackRegistries = generator.addProvider(true, new DatapackRegistryProvider(output, lookupProvider));
         generator.addProvider(true, new BiomeTagProvider(output, datapackRegistries.getRegistryProvider()));
         generator.addProvider(true, new NTDataMapProvider(output, lookupProvider));
-        generator.addProvider(true, new BacteriaPresetProvider(output));
+        generator.addProvider(true, new BacteriaMaterialProvider(output, lookupProvider));
         generator.addProvider(true, new LithostitchedInjectorProvider(output));
 
         if (ModList.get().isLoaded("modonomicon")) {