+++ src/main/java/com/breakinblocks/nautec/utils/BacteriaHelper.java
@@ -15,6 +15,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.ItemStack;
 
 import java.util.List;
@@ -56,7 +57,8 @@ public final class BacteriaHelper {
     public static Component resourceTooltip(ResourceKey<Bacteria> key) {
         Bacteria bacteria = getBacteria(Minecraft.getInstance().level.registryAccess(), key);
         MutableComponent component = Component.literal("  Resource: ").withStyle(ChatFormatting.YELLOW);
-        if (bacteria.resource() instanceof Bacteria.Resource.ItemResource(Item item)) component.append(
+        Item item = bacteria.resource().resolve();
+        if (item != Items.AIR) component.append(
                 item.getName(ItemStack.EMPTY).copy().withStyle(ChatFormatting.WHITE)
         );